driver can start new trip and go to next drop location even if he hasn't...

Closes TAC-1353

See merge request tachyonhub/tachyon-core!560

+++ aspnet-core/src/TACHYON.Application/Tracking/TrackingAppService.cs
@@ -124,7 +124,7 @@ namespace TACHYON.Tracking
             await _shippingRequestsTripManager.Start(new ShippingRequestTripDriverStartInputDto { Id = id });
         }
 
-        public async Task ChangeStatus(long id)
+        public async Task ChangeStatus(int id)
         {
             CheckIfCanAccessService(true, AppFeatures.TachyonDealer, AppFeatures.Carrier);
             await _shippingRequestsTripManager.ChangeStatus(id);
